improve error handling
- Create PopUp class to print error messages of all kinds.

- Fix Window::DestroyObject function.

- Add checks to Graphics setter functions.

- Change Window::ProcessMessages return type to uint64_t.

+++ Source/Application.cpp
@@ -21,7 +21,7 @@ void Application::DestroyObject() noexcept
 }
 
 
-int32_t Application::StartGameLoop() noexcept
+uint64_t Application::StartGameLoop() noexcept
 {
 	while (true)
 	{
